switch default bundler strategy to webpack and make bundlers pluggable
This PR makes per-module bundling strategies pluggable, and switches the default strategy to webpack.

For the set of use cases I've tried, webpack has a lot less hiccups than Rollup, primarily because (1) it's configured out of the box to deal with more Node-isms, and (2) Rollup is based around ES modules as the lingua franca, which forces all code to be interpreted in strict mode, which tends to break the weird and wacky old libraries everybody is often importing into their apps.