[wip] conditional babel transpilation
This is a work-in-progress. I'm pausing because I realized it's not exactly the right direction -- selective transpilation is actually pretty annoying to configure, because you need to make decisions about the whole deep dependency graph, which no single addon or app is really in a good position to do. One option is to make transpilation config transitive, so that if you choose to transpile something you are importing, it and all of its deps get transpiled.