[wip]
This is moving in the right direction but more is needed. Instead of modelling the tests bundle as an entrypoint, I think we should model it as a lazy chunk on top of the app entrypoint. Because that's how the relationship between vendor.js and test-support.js functions in a classic ember app.