Infer bundle type (app vs test) by original tree if known.
Prior to these changes, we would decide what bundle a given file should
end up in based on the modules relativePath that imported it. This
worked generally for application level tests that import external
packages, as well as for many addon-test-support imports.

Unfortunately, this relative path analysis falls down in addons that
intentionally customize their `treeForAddonTestSupport` so that the
`addon-test-support/` folder contents are available **without** the
`addon-name/test-support` suffix. This is done by a number of testing
only addons, but the largest offenders are `ember-qunit` and
`@ember/test-helpers`.

This introduces a new (possibly undefined) value that can be passed into
`preprocessJs` to specify the `treeType` being transpiled. ember-cli
will update to pass this value by default, but any addons doing the sort
of overrides mentioned above are _likely_ already doing a manual
override and can easily pass in the treeType in their overrides.