Fix compatiblity with ember-cli-babel@7.26.0+.
ember-cli-babel@7.24 fixed a bug where the `buildBabelOptions` result
included a number of non-babel specific properties. Unfortunately,
while we were explicitly working around that bug (the `delete`'s in the
snippet here) but we **also** ended up relying on it with the usage of
`filterExtensions` property.

In order to make it possible for ember-auto-import to have a path
forward, ember-cli-babel@7.26.0 introduces a new public API to get a
list of the supported extensions (`babelAddon.getSupportedExtensions`).

This PR uses the presence of the `getSupportedExtensions` function to
decide if it needs to clear the abnormal properties or not, and fixes
usage of ember-auto-import along with ember-cli-babel@7.26+.