refactoring away enhanced-resolve
Previously, we would resolve entrypoints to hand to webpack, and then it would do all the interior resolving. We had to resolve the entrypoints because each import may need to be resolved relative to a different package, none of which will be resolvable form down inside a broccoli pipeline.

But this has become trickier, especially now that we also want to be able to use `exports` in package.json.

Instead, we can limit our resolution to only the level of packages: we figure out which package (if any) is being targeted by a given import, and then we craft appropriate symlinks in our staging area so that that package will be visible to webpack, and then we leave all the module-scale resolutio up to webpack.