get default publicAssetURLfrom rootURL
Before if you didn't set publicAssetURL we would wait until runtime to try to guess it. Now instead we infer it from ember-cli's rootURL.

This frees the runtime __webpack_public_path__ for use by apps who really want to dynamically redirect it, and it simplifies some of our insertion code.