Avoid EBADF on ReadStream early close in Node 12
On Node <= 14, there seems to be a potential for EBADF to get raised
after a ReadStream has been closed (https://github.com/nodejs/node/issues/25335)

We can workaround this causing problems by finishing the deserializer's operations
rather than calling 'destroy' on the ReadStream mid-read. This way the Promise gets
resolved with the result before a potential EBADF can occur.

Then we can simply destroy the ReadStream explicitly when we know we're done with it.