Add babel/core to app templates
Newer versions of ember-source use peerDependencies. But because they also depend on a babel plugin with an unsatisfied peerDep on `@babel/core`, scenario-tester and fixturify-project have no way of knowing that babel/core is needed at the top-level.

This works around the problem by having babel/core in the test apps.