Add support for node type=module
Adds support for Ember addons that opt-in to node's ES module syntax via
type=module.

When writing to a directory whose package.json is controlled by a third
party, maximum compatibility is achieved by writing either `.cjs` or `.mjs`
files as appropriate, and not `.js` files.

If neither package is concerned with backwards compatibility, writing ES
modules into a `.js` file can be fine, but writing commonjs modules into
a `.js` file in a directory whose package.json is controlled by a third
party (e.g. another package) is not generally going to be
forwards-compatible

Co-authored-by: Robert Jackson <rjackson@linkedin.com>